Move coverage writer into agent and server
