+++ pkg/cli/cmds/cover_default.go
@@ -1,0 +1,7 @@
+//go:build !linux || !cover
+
+package cmds
+
+import "context"
+
+func WriteCoverage(ctx context.Context) {}
