--- pkg/server/no_cover.go
@@ -1,7 +1,0 @@
-//go:build !cover
-
-package server
-
-import "context"
-
-func writeCoverage(ctx context.Context) {}
